[VGNWB-251] added amphora utils, fixed tokenize and fixed httpreponse
