+++ annotations/annotators.py
@@ -54,7 +54,7 @@ by :mod:`annotations.views.relationtemplate_views` should be used to generate
 import requests
 from django.shortcuts import get_object_or_404, render
 from django.http import Http404
-from annotations.tasks import tokenize
+from annotations.utils import tokenize
 from annotations.utils import basepath
 from annotations.models import TextCollection, VogonUserDefaultProject
 from urlparse import urlparse
