+++ annotations/templates/annotations/upload_status.html
@@ -38,7 +38,7 @@
                                             </div>
                                         </div>
                                         <div v-else>
-                                            <div class="progress-bar progress-bar-info progress-bar-striped active" role="progressbar" aria-valuenow="80" aria-valuemin="0" aria-valuemax="100" :style="`width: ${task.percent}%`">
+                                            <div class="progress-bar progress-bar-info progress-bar-striped" role="progressbar" aria-valuenow="80" aria-valuemin="0" aria-valuemax="100" :style="`width: ${task.percent}%`">
                                                 <span class="sr-only">Complete (success)</span>
                                             </div>
                                         </div>
