+++ annotations/utils.py
@@ -34,3 +34,23 @@ def basepath(request):
     else:
         scheme = 'http://'
     return scheme + request.get_host() + settings.SUBPATH
+
+
+def tokenize(content, delimiter=u' '):
+    """
+    In order to annotate a text, we must first wrap "annotatable" tokens
+    in <word></word> tags, with arbitrary IDs.
+
+    Parameters
+    ----------
+    content : unicode
+    delimiter : unicode
+        Character or sequence by which to split and join tokens.
+
+    Returns
+    -------
+    tokenizedContent : unicode
+    """
+    chunks = content.split(delimiter)
+    pattern = u'<word id="{0}">{1}</word>'
+    return delimiter.join([pattern.format(i, c) for i, c in enumerate(chunks)])
